pyproject.toml
[project]
name = "twin-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "shinkeonkim",email = "dev.shinkeonkim@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "chromadb (>=1.0.11,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "slack-bolt (>=1.23.0,<2.0.0)",
    "slack-sdk (>=3.35.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-community (>=0.3.25,<0.4.0)",
    "langchain-core (>=0.3.65,<0.4.0)",
    "langchain-openai (>=0.3.23,<0.4.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "drf-spectacular-sidecar (>=2025.6.1,<2026.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "django-extensions (>=4.1,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.0"
pre-commit = "^4.0.1"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.9.3"
mypy = "^1.8.0"
autoflake = "^2.2.0"

[tool.poetry]
name = "twin-ai"
package-mode = false
